Consolidate keyword-demotion transforms into one module
The five xform_demote_* passes had grown organically, each an
almost-identical O(n) walk that rewrites a set of keyword tokens to
Identifier when the gating flag is off. The keyword-to-flag mapping was
scattered across three places per keyword (token.rs doc-comment, the
apply() body, and options.rs), with no single source of truth, and each
new keyword required a new module plus a registration line in
tokenize_program.

Collapse the flag-gated demoters into one xform_demote_keywords module
whose single match on token_type is the one place that defines which
keyword demotes under which flag. The context-sensitive TIME case (which
depends on neighbouring tokens and has inverted polarity) stays a private
apply_time helper in the same module, called from apply. Behaviour is
preserved: the five passes targeted disjoint token sets and TIME only
inspects neighbour kinds that are never demotion targets, so ordering
never mattered. All unit tests from the five modules are ported verbatim.

Update the syntax-support-guide steering doc so the documented workflow is
"add a match arm" instead of "add a new xform module and register it".
