+++ compiler/parser/src/lib.rs
@@ -14,11 +14,7 @@ mod rule_token_no_partial_access_syntax;
 mod vars;
 mod xform_assign_file_id;
 mod xform_collapse_pragmas;
-mod xform_demote_edition3_keywords;
-mod xform_demote_oop_keywords;
-mod xform_demote_reference_keyword;
-mod xform_demote_short_circuit_operators;
-mod xform_demote_time_keyword;
+mod xform_demote_keywords;
 mod xform_tokens;
 
 use crate::parser::{parse_library, parse_statements};
@@ -65,11 +61,7 @@ pub fn tokenize_program(
 
     let tokens = xform_collapse_pragmas::apply(tokens, options);
     let mut tokens = insert_keyword_statement_terminators(tokens, file_id, options);
-    xform_demote_edition3_keywords::apply(&mut tokens, options);
-    xform_demote_reference_keyword::apply(&mut tokens, options);
-    xform_demote_time_keyword::apply(&mut tokens, options);
-    xform_demote_oop_keywords::apply(&mut tokens, options);
-    xform_demote_short_circuit_operators::apply(&mut tokens, options);
+    xform_demote_keywords::apply(&mut tokens, options);
     let result = check_tokens(&tokens, options);
     match result {
         Ok(_) => {}
